Kill VersionComparator
